+++ src/Exception/UserAlreadyMappedToApplication.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Adrenth\Raindrop\Exception;
+
+/**
+ * Class UserAlreadyMappedToApplication
+ *
+ * @package Adrenth\Raindrop\Exception
+ */
+class UserAlreadyMappedToApplication extends RegisterUserFailed
+{
+
+}