+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Hydro Raindrop SDK PHP
 
+## 1.1.0
+
+* Added `UserAlreadyMappedToApplication` and `UsernameDoesNotExist` exceptions to allow custom error handling.
+* Change `User-Agent` header value. Example: `adrenth.raindrop-sdk (PHP) version 1.1.0`
+
 ## 1.0.0
+
 * Initial release.